Question 12 Lifecycle hook experiment
12) Make it so a random person is generated when the page first loads

the vue data object seems to load between beforeCreate() and created() lifecycle hooks.  checking whether possible to init empty data props and dynamically update with randomly generated values on the beforeCreate() lifecycle.

Nope:  Methods are not loaded at that point in the lifecycle error: this.randomPerson is not a function